refactor file checks in its own file
