+++ tasks/quality_gates.py
@@ -1,9 +1,6 @@
-import glob
 import os
 import random
 import re
-import sys
-import tempfile
 import traceback
 import typing
 from collections.abc import Callable
@@ -24,21 +21,11 @@ from tasks.libs.common.git import (
 )
 from tasks.libs.common.utils import running_in_ci
 from tasks.libs.package.size import InfraError
-from tasks.static_quality_gates.experimental_gates import FileInfo
-from tasks.static_quality_gates.experimental_gates import (
-    compare_inventory as _compare_inventory,
-)
-from tasks.static_quality_gates.experimental_gates import (
-    inventory_changes_to_comment as _inventory_changes_to_comment,
-)
 from tasks.static_quality_gates.experimental_gates import (
     measure_image_local as _measure_image_local,
 )
 from tasks.static_quality_gates.experimental_gates import (
     measure_package_local as _measure_package_local,
-)
-from tasks.static_quality_gates.experimental_gates import (
-    print_inventory_diff as _print_inventory_diff,
 )
 from tasks.static_quality_gates.gates import (
     GateMetricHandler,
@@ -1101,84 +1088,3 @@ def measure_image_local(
         include_layer_analysis=include_layer_analysis,
         debug=debug,
     )
-
-
-@task
-def compare_inventories(ctx, parent_inventory_report, current_inventory_report):
-    with open(parent_inventory_report) as f:
-        parent_inventory = yaml.safe_load(f)
-    with open(current_inventory_report) as f:
-        current_inventory = yaml.safe_load(f)
-    parent_file_inventory = [FileInfo(**values) for values in parent_inventory['file_inventory']]
-    current_file_inventory = [FileInfo(**values) for values in current_inventory['file_inventory']]
-    added, removed, changed = _compare_inventory(parent_file_inventory, current_file_inventory)
-    if len(added) == 0 and len(removed) == 0 and len(changed) == 0:
-        print(color_message('✅ No change detected', 'green'))
-        body = "No change detected"
-    else:
-        _print_inventory_diff(added, removed, changed)
-        body = _inventory_changes_to_comment(added, removed, changed)
-    return body
-
-
-def _get_parent_report(ctx, parent_sha, gate_name: str, output: str):
-    """
-    Fetch the quality gates report from the base commit.
-    """
-    aws_cmd = "aws.exe" if sys.platform == 'win32' else "aws"
-    s3_url = f"s3://dd-ci-artefacts-build-stable/datadog-agent/static_quality_gates/GATE_REPORTS/{parent_sha}/{gate_name}_size_report_{parent_sha[:8]}.yml"
-    ctx.run(f"{aws_cmd} s3 cp --only-show-errors {s3_url} {output}", warn=True)
-
-
-def _filter_files(path: str) -> bool:
-    pipeline_id = os.environ.get('CI_PIPELINE_ID')
-    # Avoid files such as opt/datadog-packages/run/datadog-agent/7.76.0-devel.git.716.8d5ec09.pipeline.91823583
-    if pipeline_id and f'pipeline.{pipeline_id}' in path:
-        return False
-    return path not in [
-        # Installer symlinks which we don't want to include
-        'opt/datadog-packages/datadog-agent/experiment',
-        'opt/datadog-packages/datadog-agent/stable',
-    ]
-
-
-@task
-def check_files(ctx, branch_name, reports_folder):
-    package_types = ['deb', 'rpm']
-    parent_sha = get_ancestor(ctx, branch_name)
-    pr_comment = f"File checks results against ancestor [{parent_sha[:8]}](https://github.com/DataDog/datadog-agent/commit/{parent_sha}):\n\n"
-
-    for package_type in package_types:
-        for artifact in glob.glob(f'{reports_folder}/datadog-agent*.{package_type}'):
-            # deb pattern is $packagename-$version
-            # rpm pattern is $packagename_$version
-            if '-dbg-' in artifact or '-dbg_' in artifact:
-                continue
-            pr_comment += f'## Results for {os.path.basename(artifact)}:\n'
-            arch = "amd64"
-            if 'aarch64' in artifact or 'arm64' in artifact:
-                arch = "arm64"
-            gate_short_name = f'agent_{package_type}_{arch}'
-            report_filename = f'{gate_short_name}_{arch}_size_report_{os.environ["CI_COMMIT_SHORT_SHA"]}.yml'
-            measure_package_local(
-                ctx,
-                artifact,
-                f'static_quality_gate_{gate_short_name}',
-                output_path=report_filename,
-                build_job_name=os.environ['CI_JOB_NAME'],
-                debug=True,
-                filter=_filter_files,
-            )
-            # Upload the report to S3
-            bucket_base_path = "s3://dd-ci-artefacts-build-stable/datadog-agent/static_quality_gates/GATE_REPORTS/"
-            ctx.run(
-                f'aws s3 cp --only-show-errors --region us-east-1 --sse AES256 {report_filename} {bucket_base_path}/{report_filename}'
-            )
-
-            parent_report_file = tempfile.NamedTemporaryFile()
-            _get_parent_report(ctx, parent_sha, gate_short_name, parent_report_file.file.name)
-            body = compare_inventories(ctx, parent_report_file.file.name, report_filename)
-            pr_comment += body
-
-    pr = get_pr_for_branch(branch_name)
-    pr_commenter(ctx, title='File checks summary', body=pr_comment, pr=pr)
