+++ tasks/static_quality_gates/experimental_gates.py
@@ -22,7 +22,6 @@ from tasks.libs.common.color import color_message
 from tasks.libs.package.size import extract_package, file_size
 from tasks.static_quality_gates.gates import (
     QualityGateConfig,
-    byte_to_string,
     create_quality_gate_config,
 )
 
@@ -1401,97 +1400,3 @@ def measure_image_local(
     except Exception as e:
         print(color_message(f"❌ Image measurement failed: {e}", "red"))
         raise
-
-
-def compare_inventory(
-    previous_inventory: list[FileInfo], current_inventory: list[FileInfo]
-) -> tuple[list[FileInfo], list[FileInfo], dict[str, FileChange]]:
-    removed_files = []
-    added_files = []
-    changed_files = {}
-
-    previous_files = {info.relative_path: info for info in previous_inventory}
-    current_files = {info.relative_path: info for info in current_inventory}
-    for path, previous in previous_files.items():
-        if path not in current_files:
-            removed_files.append(previous)
-            continue
-        current = current_files[path]
-        size_change = previous.size_bytes - current.size_bytes
-        changed_flags = FileChange.Flags(0)
-        size_percent = None
-        if size_change:
-            size_percent = (current.size_bytes - previous.size_bytes) / previous.size_bytes * 100
-            if size_percent > 10:
-                changed_flags |= FileChange.Flags.Size
-        if current.chmod != previous.chmod:
-            changed_flags |= FileChange.Flags.Permissions
-        if current.owner != previous.owner:
-            changed_flags |= FileChange.Flags.Owner
-        if current.group != previous.group:
-            changed_flags |= FileChange.Flags.Group
-
-        if changed_flags:
-            changed_files[path] = FileChange(
-                flags=changed_flags, previous=previous, current=current, size_percent=size_percent
-            )
-        # Remove entries that were present in both parent & current so that when we're done
-        # the current list only contains new files
-        del current_files[path]
-    added_files = list(current_files.values())
-    return added_files, removed_files, changed_files
-
-
-def _display_change_summary(change: FileChange):
-    print(color_message(f'Summary of changes to {change.current.relative_path}', "orange"))
-    if change.flags & FileChange.Flags.Permissions:
-        print(f'    Permission changed: {oct(change.previous.chmod)} -> {oct(change.current.chmod)}')
-    if change.flags & FileChange.Flags.Size:
-        color = "red" if change.size_percent > 0 else "green"
-        change_str = color_message(f'{change.size_percent:.2f}%', color)
-        print(
-            f'    Size changed by {change_str} ({byte_to_string(change.previous.size_bytes)} -> {byte_to_string(change.current.size_bytes)})'
-        )
-    if change.flags & (FileChange.Flags.Owner | FileChange.Flags.Group):
-        print(
-            f'    File owner/group changed: {change.previous.owner}:{change.previous.group} -> {change.current.owner}:{change.current.group}'
-        )
-
-
-def print_inventory_diff(added, removed, changed):
-    if len(added) > 0:
-        print(color_message('➕ New files added:', "orange"))
-        for f in added:
-            print(color_message(f'    - {f.relative_path} ({byte_to_string(f.size_bytes)})', "orange"))
-    if len(removed) > 0:
-        print(color_message('❌ Old files removed:', "orange"))
-        for f in removed:
-            print(color_message(f'    - {f.relative_path} ({byte_to_string(f.size_bytes)})', "orange"))
-    if len(changed) > 0:
-        print(color_message('⚠️ Some files modifications need review:', "orange"))
-        for change in changed.values():
-            _display_change_summary(change)
-
-
-def inventory_changes_to_comment(added, removed, changed):
-    body = "## Detected file changes:\n"
-    if len(added):
-        body += "### Added files:\n"
-        for f in added:
-            body += f"* `{f.relative_path}` ({byte_to_string(f.size_bytes)})\n"
-    if len(removed):
-        body += "### Removed files:\n"
-        for f in removed:
-            body += f"* `{f.relative_path}` ({byte_to_string(f.size_bytes)})\n"
-    if len(changed):
-        body += "### Changed files:\n"
-        for path, change in changed.items():
-            change_str = f"* `{path}`:\n"
-            if change.flags & FileChange.Flags.Permissions:
-                change_str += f"  * Permission changed: {oct(change.previous.chmod)} -> {oct(change.current.chmod)}"
-            if change.flags & FileChange.Flags.Size:
-                change_str += f'  * Size changed: {change.size_percent:+.2f}% ({byte_to_string(change.previous.size_bytes)} -> {byte_to_string(change.current.size_bytes)})\n'
-            if change.flags & (FileChange.Flags.Owner | FileChange.Flags.Group):
-                change_str += f'  * File owner/group changed: {change.previous.owner}:{change.previous.group} -> {change.current.owner}:{change.current.group}'
-            body += change_str
-    return body
